Feat/#106/페이지 총 댓글 추가, 이벤트 페이지 시나리오 추가 (#107)
* [feat] Integer -> int 배열로 수정

* [feat] 사용되지 않은 유저에 대한 정보 0으로 처리

* [feat] 정보 저장 로직 수정

* [test] 생성자 변경에 따른 DrawControllerTest 수정

* [test] 생성자 변경에 따른 DrawServiceTest 수정

* [test] 로직 수정에 따른 테스트 코드 수정

* [feat] 해당 날짜와 응모 시나리오 조회하는 api 추가

* [feat] 해당 날짜와 응모 시나리오 조회하는 메서드 추가

* [feat] admin이 image를 처리하지 않고 보냈을 시 예외 처리

* [feat] 트랜잭션 추가

* [feat] 응모 결과 페이지 응답 시 총 댓글 개수 추가

* [feat] 해당 날짜와 응모 시나리오 조회하는 메서드를 위한 응답 클래스 생성

+++ src/main/java/softeer/team_pineapple_be/domain/draw/service/DrawRewardInfoService.java
@@ -4,11 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import softeer.team_pineapple_be.domain.draw.domain.DrawRewardInfo;
+import softeer.team_pineapple_be.domain.draw.exception.DrawErrorCode;
 import softeer.team_pineapple_be.domain.draw.repository.DrawRewardInfoRepository;
 import softeer.team_pineapple_be.domain.draw.request.DrawRewardInfoListRequest;
 import softeer.team_pineapple_be.domain.draw.response.DrawRewardInfoListResponse;
 import softeer.team_pineapple_be.global.cloud.service.S3DeleteService;
 import softeer.team_pineapple_be.global.cloud.service.S3UploadService;
+import softeer.team_pineapple_be.global.exception.RestApiException;
 
 import java.io.IOException;
 import java.util.List;
@@ -48,6 +50,13 @@ public class DrawRewardInfoService {
     public void setDrawRewardInfoList(DrawRewardInfoListRequest request) {
         List<DrawRewardInfo> rewardInfoList = request.getRewards().stream()
                 .map(rewardInfoRequest -> {
+                    if(rewardInfoRequest.getImage() == null){
+                        DrawRewardInfo drawRewardInfo = drawRewardInfoRepository.findById(rewardInfoRequest.getRanking()).orElseThrow(() -> new RestApiException(DrawErrorCode.NO_PRIZE));
+                        return new DrawRewardInfo(rewardInfoRequest.getRanking(),
+                                rewardInfoRequest.getName(),
+                                rewardInfoRequest.getStock(),
+                                drawRewardInfo.getImage());
+                    }
                     String fileName = DRAW_REWAR_INFO_FOLDER + rewardInfoRequest.getRanking() + "/";
                     s3DeleteService.deleteFolder(fileName);
                     String image;
